tungnv_b23dckh130_training_gd1/NewUseCase/BackEnd/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
alembic==1.13.1
redis==7.1.0
pika==1.3.2